requirements.txt
# DRP Pipeline Dependencies
# Add dependencies here as the project develops

typer>=0.9.0
anthropic>=0.40.0
dspy-ai>=2.5.0
pytest>=7.0.0
pytest-cov>=4.0.0
playwright>=1.49.0
requests>=2.31.0
flask>=3.0.0
watchdog>=4.0.0
beautifulsoup4>=4.12.0
mcp[cli]>=1.0.0
pytest>=8.0.0

# Optional: improves catalog.data.gov fetch (Chrome TLS impersonation, may bypass AWS WAF)
# curl_cffi>=0.6.0

# Optional: for publisher Google Sheet update (master inventory)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
openpyxl>=3.1.0
